websockets/constructor/021.html

<!doctype html>
<title>WebSockets: Same sub protocol twice</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
test(function() {
	assert_throws("SyntaxError", function(){
		new WebSocket(SCHEME_DOMAIN_PORT+"/ws/protocol_array",["foobar", "foobar"])
	})
});
</script>
