Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/constructor/021-expected.txt

CONSOLE MESSAGE: line 10: Wrong protocol for WebSocket 'foobar, foobar'

PASS WebSockets: Same sub protocol twice 
